force update for CPU only mode

+++ src/force_update.bat
@@ -3,7 +3,12 @@
 cd %~dp0\..
 .\git\bin\git.exe reset --hard
 .\git\bin\git.exe pull --rebase
-.\python\python.exe -m pip install -U -r .\src\requirements.txt
+if exist ../python/CPU (
+	.\python\python.exe -m pip install -U -r .\src\requirements.cpu.txt
+) else (
+	.\python\python.exe -m pip install -U -r .\src\requirements.txt
+)
+
 
 echo Done! Press any key to exit...
 pause >nul
